# backend/openapi.yaml
openapi: 3.1.0
info:
  title: BudgetBuddy API
  version: "1"
  description: |
    Operational header behavior: responses include `X-Request-Id`.
    If request provides `X-Request-Id`, the API echoes it; otherwise it generates one.
servers:
- url: /api
security:
- BearerAuth: []
paths:
...
  /transactions:
    get:
      summary: List transactions (ordered by date desc, then created_at desc)
      description: |
        Filters `type`, `account_id`, `category_id`, `from`, `to`, and `include_archived` are combined as AND predicates.
        Cursor pagination uses an opaque `base64url(JSON)` token and deterministic ordering by `date DESC, created_at DESC, id DESC`.
        For stable datasets, paging is deterministic with no overlaps or skips; terminal page returns `next_cursor: null`.
        Under concurrent mutations, pagination is best-effort and does not guarantee snapshot semantics.
        Compatibility policy: `GET /transactions` accepts legacy cursors containing only `date` and `id` (without `created_at`) for backward compatibility in API v1.
        `GET /accounts` and `GET /categories` require current cursor fields only; malformed or unsupported cursor shapes return canonical `400 Invalid cursor`.
      parameters:
      - name: include_archived
        in: query
        required: false
        schema:
          type: boolean
          default: false
      - name: type
        in: query
        required: false
        schema:
          type: string
          enum: [ income, expense ]
      - name: account_id
        in: query
        required: false
        schema:
          type: string
          format: uuid
      - name: category_id
        in: query
        required: false
        schema:
          type: string
          format: uuid
      - name: from
        in: query
        required: false
        schema:
          type: string
          format: date
      - name: to
        in: query
        required: false
        schema:
          type: string
          format: date
      - name: cursor
        in: query
        required: false
        description: Opaque `base64url(JSON)` token built from `date`, `created_at`, and `id` of the last item in the previous page.
        schema:
          type: string
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          default: 20
          minimum: 1
          maximum: 100
      responses:
        "200":
          description: OK
          headers:
            X-Request-Id:
              $ref: "#/components/headers/X-Request-Id"
          content:
            application/vnd.budgetbuddy.v1+json:
              schema:
                type: object
                required: [ items, next_cursor ]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/Transaction"
                  next_cursor:
                    type: [ string, "null" ]

        "400":
          description: Invalid cursor or invalid date range (from must be <= to)
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "401":
          description: Unauthorized
          headers:
            X-Request-Id:
              $ref: "#/components/headers/X-Request-Id"
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "406":
          description: Not Acceptable
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"

    post:
      summary: Create transaction
      requestBody:
        required: true
        content:
          application/vnd.budgetbuddy.v1+json:
            schema:
              $ref: "#/components/schemas/TransactionCreate"
      responses:
        "201":
          description: Created
          content:
            application/vnd.budgetbuddy.v1+json:
              schema:
                $ref: "#/components/schemas/Transaction"
        "400":
          description: Invalid request
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "401":
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "403":
          description: Forbidden (resource is not owned by authenticated user)
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "409":
          description: Business rule conflict (e.g. account archived, category archived, type mismatch)
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "406":
          description: Not Acceptable
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
  /transactions/import:
    post:
      summary: Import transactions in bulk
      requestBody:
        required: true
        content:
          application/vnd.budgetbuddy.v1+json:
            schema:
              $ref: "#/components/schemas/TransactionImportRequest"
      responses:
        "200":
          description: Import result
          content:
            application/vnd.budgetbuddy.v1+json:
              schema:
                $ref: "#/components/schemas/TransactionImportResult"
        "400":
          description: Invalid request (payload validation or import batch limit exceeded)
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "401":
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "403":
          description: Forbidden (resource is not owned by authenticated user)
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "406":
          description: Not Acceptable
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
  /transactions/export:
    get:
      summary: Export transactions as CSV
      parameters:
      - name: type
        in: query
        required: false
        schema:
          type: string
          enum: [ income, expense ]
      - name: account_id
        in: query
        required: false
        schema:
          type: string
          format: uuid
      - name: category_id
        in: query
        required: false
        schema:
          type: string
          format: uuid
      - name: from
        in: query
        required: false
        schema:
          type: string
          format: date
      - name: to
        in: query
        required: false
        schema:
          type: string
          format: date
      responses:
        "200":
          description: CSV export stream
          content:
            text/csv:
              schema:
                type: string
                format: binary
        "400":
          description: Invalid request (for example invalid date range)
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "401":
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "429":
          description: Too Many Requests
          headers:
            Retry-After:
              $ref: "#/components/headers/Retry-After"
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "406":
          description: Not Acceptable
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
  /transactions/{transaction_id}:
    parameters:
    - name: transaction_id
      in: path
      required: true
      schema:
        type: string
        format: uuid

    get:
      summary: Get transaction
      responses:
        "200":
          description: OK
          content:
            application/vnd.budgetbuddy.v1+json:
              schema:
                $ref: "#/components/schemas/Transaction"
        "400":
          description: Invalid request
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "401":
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "403":
          description: Forbidden (resource is not owned by authenticated user)
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "406":
          description: Not Acceptable
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"

    patch:
      summary: Update or restore transaction
      requestBody:
        required: true
        content:
          application/vnd.budgetbuddy.v1+json:
            schema:
              $ref: "#/components/schemas/TransactionUpdate"
      responses:
        "200":
          description: Updated (or restored when archived_at=null)
          content:
            application/vnd.budgetbuddy.v1+json:
              schema:
                $ref: "#/components/schemas/Transaction"
        "400":
          description: Invalid request
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "401":
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "403":
          description: Forbidden (resource is not owned by authenticated user)
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "409":
          description: Business rule conflict (e.g. account archived, category archived, type mismatch)
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "406":
          description: Not Acceptable
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"

    delete:
      summary: Archive transaction
      responses:
        "204":
          description: No Content
        "400":
          description: Invalid request
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "401":
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "403":
          description: Forbidden (resource is not owned by authenticated user)
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
        "406":
          description: Not Acceptable
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
...
components:
  headers:
    X-Request-Id:
      description: Correlation id for tracing requests. Propagated when provided by client, generated otherwise.
      schema:
        type: string
    Retry-After:
      description: Seconds to wait before retrying after rate limiting.
      schema:
        type: integer
        minimum: 1
...
  schemas:
    ProblemDetails:
      type: object
      required: [ type, title, status ]
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
          format: uri
...
    Transaction:
      type: object
      required: [ id, type, account_id, category_id, amount_cents, date, archived_at, created_at, updated_at ]
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
          enum: [ income, expense ]
        account_id:
          type: string
          format: uuid
        category_id:
          type: string
          format: uuid
        amount_cents:
          type: integer
          minimum: 1
          maximum: 999999999999
        date:
          type: string
          format: date
        merchant:
          type: [ string, "null" ]
        note:
          type: [ string, "null" ]
        archived_at:
          type: [ string, "null" ]
          format: date-time
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    TransactionCreate:
      type: object
      required: [ type, account_id, category_id, amount_cents, date ]
      properties:
        type:
          type: string
          enum: [ income, expense ]
        account_id:
          type: string
          format: uuid
        category_id:
          type: string
          format: uuid
        amount_cents:
          type: integer
          minimum: 1
          maximum: 999999999999
        date:
          type: string
          format: date
        merchant:
          type: string
        note:
          type: string
    TransactionUpdate:
      type: object
      properties:
        type:
          type: string
          enum: [ income, expense ]
        account_id:
          type: string
          format: uuid
        category_id:
          type: string
          format: uuid
        amount_cents:
          type: integer
          minimum: 1
          maximum: 999999999999
        date:
          type: string
          format: date
        merchant:
          type: [ string, "null" ]
        note:
          type: [ string, "null" ]
        archived_at:
          type: [ string, "null" ]
          format: date-time
    TransactionImportFailure:
      type: object
      required: [ index, message ]
      properties:
        index:
          type: integer
          minimum: 0
        message:
          type: string
        problem:
          $ref: "#/components/schemas/ProblemDetails"
    TransactionImportResult:
      type: object
      required: [ created_count, failed_count, failures ]
      properties:
        created_count:
          type: integer
          minimum: 0
        failed_count:
          type: integer
          minimum: 0
        failures:
          type: array
          items:
            $ref: "#/components/schemas/TransactionImportFailure"
    TransactionImportRequest:
      type: object
      required: [ mode, items ]
      properties:
        mode:
          type: string
          enum: [ all_or_nothing, partial ]
          default: partial
        items:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/TransactionCreate"
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT